requirements.txt
asgiref==3.3.4
certifi==2021.5.30
cffi==1.14.5
chardet==4.0.0
cryptography==3.4.7
Django==3.2.4
django-rest-knox==4.1.0
djangorestframework==3.12.4
idna==2.10
Pillow==8.2.0
pycparser==2.20
python-decouple==3.4
pytz==2021.1
requests==2.25.1
sqlparse==0.4.1
typing-extensions==3.10.0.0
urllib3==1.26.5
